added session capabilites and a logout function

+++ src/main/java/com/JohnHaney/OpenJob/DAO/UserRepoIF.java
@@ -7,6 +7,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import com.JohnHaney.OpenJob.models.User;
 
 
-public interface UserRepoIF extends JpaRepository<User, Integer>{
+public interface UserRepoIF extends JpaRepository<User, Long>{
 	Optional<User> findByUsername(String username);
 }
